Move compiler passes one level up
