+++ src/PurgatoryBundle.php
@@ -4,10 +4,10 @@ declare(strict_types=1);
 
 namespace Sofascore\PurgatoryBundle;
 
-use Sofascore\PurgatoryBundle\DependencyInjection\CompilerPass\ControllerClassMapPass;
-use Sofascore\PurgatoryBundle\DependencyInjection\CompilerPass\RegisterExpressionLanguageProvidersPass;
-use Sofascore\PurgatoryBundle\DependencyInjection\CompilerPass\RegisterPurgerPass;
-use Sofascore\PurgatoryBundle\DependencyInjection\CompilerPass\RegisterRouteParamServicesPass;
+use Sofascore\PurgatoryBundle\DependencyInjection\ControllerClassMapCompilerPass;
+use Sofascore\PurgatoryBundle\DependencyInjection\RegisterExpressionLanguageProvidersCompilerPass;
+use Sofascore\PurgatoryBundle\DependencyInjection\RegisterPurgerCompilerPass;
+use Sofascore\PurgatoryBundle\DependencyInjection\RegisterRouteParamServicesCompilerPass;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
@@ -15,10 +15,10 @@ final class PurgatoryBundle extends Bundle
 {
     public function build(ContainerBuilder $container): void
     {
-        $container->addCompilerPass(new ControllerClassMapPass());
-        $container->addCompilerPass(new RegisterExpressionLanguageProvidersPass());
-        $container->addCompilerPass(new RegisterPurgerPass());
-        $container->addCompilerPass(new RegisterRouteParamServicesPass());
+        $container->addCompilerPass(new ControllerClassMapCompilerPass());
+        $container->addCompilerPass(new RegisterExpressionLanguageProvidersCompilerPass());
+        $container->addCompilerPass(new RegisterPurgerCompilerPass());
+        $container->addCompilerPass(new RegisterRouteParamServicesCompilerPass());
     }
 
     public function getPath(): string
